fix: Deny removal of admin role

+++ mcm-backend/mcm-core/src/main/java/at/ac/tuwien/model/change/management/core/service/UserRoleServiceImpl.java
@@ -81,6 +81,9 @@ public class UserRoleServiceImpl implements UserRoleService {
         if(!userRoleRepository.existsById(name)) {
             throw new UserRoleNotFoundException("Role %s not found".formatted(name));
         }
+        if(name.equals("admin")) {
+            throw new IllegalArgumentException("Cannot delete admin role");
+        }
         userRoleRepository.deleteById(name);
     }
 }
